+++ packages/shield-controller/src/polling-with-policy.test.ts
@@ -1,6 +1,7 @@
+import { HttpError } from '@metamask/controller-utils';
+
 import { PollingWithCockatielPolicy } from './polling-with-policy';
 import { delay } from '../tests/utils';
-import { HttpError } from '@metamask/controller-utils';
 
 describe('PollingWithCockatielPolicy', () => {
   it('should return the success result', async () => {
@@ -22,7 +23,8 @@ describe('PollingWithCockatielPolicy', () => {
           setTimeout(() => {
             // eslint-disable-next-line jest/no-conditional-in-test
             if (invocationCount < 3) {
-              reject(new HttpError(412));
+              // eslint-disable-next-line @typescript-eslint/prefer-promise-reject-errors -- HttpError is a valid error which extends Error class
+              reject(new HttpError(412, 'Results are not available yet'));
             }
             resolve('test');
           }, 100);
@@ -39,6 +41,7 @@ describe('PollingWithCockatielPolicy', () => {
       .fn()
       .mockImplementation(async (_abortSignal: AbortSignal) => {
         return new Promise((_resolve, reject) => {
+          // eslint-disable-next-line @typescript-eslint/prefer-promise-reject-errors -- HttpError is a valid error which extends Error class
           reject(new HttpError(500, 'Internal server error'));
         });
       });
@@ -58,6 +61,7 @@ describe('PollingWithCockatielPolicy', () => {
       .mockImplementation(async (_abortSignal: AbortSignal) => {
         return new Promise((_resolve, reject) => {
           setTimeout(() => {
+            // eslint-disable-next-line @typescript-eslint/prefer-promise-reject-errors -- HttpError is a valid error which extends Error class
             reject(new HttpError(412, 'Results are not available yet'));
           }, 100);
         });